feat: move wasm out of experimental and incentive into experimental
